added business logic

+++ src/main/java/org/example/task8/dto/ConvertCurrencyResponse.java
@@ -1,8 +1,10 @@
 package org.example.task8.dto;
 
+import java.math.BigDecimal;
+
 public record ConvertCurrencyResponse(
         String fromCurrency,
         String toCurrency,
-        double convertedAmount
+        BigDecimal convertedAmount
 ) {
 }
